+++ src/cqspoolterrain/__init__.py
@@ -1,3 +1,5 @@
 from .Base import Base
 from .Spool import Spool
 from .Cradle import Cradle
+from .StairLift import StairLift
+from .ControlPlatform import ControlPlatform