+++ changes.md
@@ -1,5 +1,10 @@
 # CQSpoolTerrain changelog
 
 ## Main wip
+## 0.0.2
+* Upped cqterrain verison 0.1.8
+* Made PowerStation class.
+
+## 0.0.1
 * Added Initial Spool Mockup.
-* Added Initial Cradle and example
+* Added Initial Cradle and example.